Finacials/requirements.txt
# PDF to Excel/CSV Converter Requirements

# Core Flask Framework
Flask==2.3.3
Werkzeug==2.3.7

# PDF Processing
PyMuPDF==1.23.5
pdf2image==1.16.3
PyPDF2==3.0.1

# Image Processing
Pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.3

# OCR (Optical Character Recognition)
pytesseract==0.3.10

# Data Processing
pandas==2.1.0
openpyxl==3.1.2

# File Handling
chardet==5.2.0

# Note: uuid, threading, json, re, os, logging, datetime are built-in Python modules
